Projeler/YouTube_Kapak/requirements.txt
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-genai==1.0.0
google-generativeai==0.7.0
python-dotenv==1.0.1
requests==2.31.0
